oracle responses and get presales endpoint

+++ models/set_user_claim_response.go
@@ -1,6 +1,6 @@
 package models
 
 type SetUserClaimResponse struct {
-	Status        uint `json:"status"`
-	ClaimedAmount uint `json:"claimed_amount"`
+	Status        uint    `json:"status"`
+	ClaimedAmount float64 `json:"claimed_amount"`
 }
